feat: /quiz -> /game/quiz
