+++ apps/webview/public/assets/game/quiz/cursor-choiced.webp
