+++ apps/webview/public/assets/game/quiz/cursor-unchoiced.webp
