+++ apps/webview/public/assets/game/quiz/ox_x.webp
