+++ apps/webview/public/assets/game/quiz/quiz-cat.webp
