+++ apps/webview/src/app/[locale]/game/quiz/_components/CreateOrSolve/QuizTypeCard.tsx
