+++ apps/webview/src/app/[locale]/game/quiz/layout.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from 'react';
 
-import { MobileLayout } from '@/app/[locale]/quiz/_components/MobileLayout';
+import { MobileLayout } from '@/app/[locale]/game/quiz/_components/MobileLayout';
 
 export default function QuizLayout({ children }: { children: React.ReactNode }) {
   return (
