+++ apps/webview/src/app/[locale]/game/quiz/solve/_components/solving/QuizProgressBar.tsx
