+++ apps/webview/src/app/[locale]/game/quiz/solve/_constants/solveQuiz.constants.ts
